Update index.html
Notice the plain text at the top of the file. Jekyll calls this the Front-matter. 
Any file on your site that contains this will be processed by Jekyll. 
Every time you commit a file that specifies layout: default at the top, Jekyll will magically generate the full HTML document - 
it will do this by replacing {{ content }} in _layouts/default.html with the contents of the committed file. Awesome!