Create post.html
We'll want to create a new layout for our blog posts called post.html and a folder to store each individual post called _posts/.

Start by creating the layout. Create a file named post.html in your _layouts folder. Notice the post layout uses the default layout as it's base, and adds a couple new liquid tags to print the title of the post and date: