Update post.html
<!-- A Jekyll-based blog uses the same conventions that we've familiarized ourselves with in the previous steps.
But takes things further by adding a few more for us to follow. Jekyll is very flexible allowing you to extend your site as you wish. 
However, in this guide we're only going to cover the basics: creating a post, making a page to list our posts, 
and creating a custom permalink for posts, and creating an RSS feed for the blog. 
Notice the post layout uses the default layout as it's base, and adds a couple new liquid tags to print the title of the post and date -->